day05 cpp

+++ C++/Day01.cpp
@@ -6,9 +6,9 @@ namespace Day01
 
 	int Part1(const std::vector<int> v, const int target)
 	{
-		for (int i = 0; i < v.size(); ++i)
+		for (size_t i = 0; i < v.size(); ++i)
 	  {
-			for (int j = i+1; j < v.size(); ++j)
+			for (size_t j = i+1; j < v.size(); ++j)
 		  {
 		  	if (v[i] + v[j] == target)
 		  	{
@@ -22,11 +22,11 @@ namespace Day01
 
 	int Part2(const std::vector<int> v, const int target)
 	{
-		for (int i = 0; i < v.size(); ++i)
+		for (size_t i = 0; i < v.size(); ++i)
 	  {
-			for (int j = i+1; j < v.size(); ++j)
+			for (size_t j = i+1; j < v.size(); ++j)
 		  {
-		  	for (int k = j+1; k < v.size(); ++k)
+		  	for (size_t k = j+1; k < v.size(); ++k)
 			  {
 			  	if (v[i] + v[j] + v[k] == target)
 			  	{
